requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
betterlogging==1.0.0
certifi==2024.12.14
charset-normalizer==3.4.3
configparser==7.1.0
environs==14.1.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
marshmallow==3.26.0
multidict==6.1.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.3.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
typing_extensions==4.12.2
urllib3==2.5.0
yarl==1.18.3
